Make TestSurface public.
