+++ painter_test.go
@@ -1,13 +1,12 @@
 package tui
 
 import (
-	"bytes"
 	"image"
 	"testing"
 )
 
 func TestMask_Full(t *testing.T) {
-	surface := newTestSurface(10, 10)
+	surface := NewTestSurface(10, 10)
 
 	p := NewPainter(surface, NewTheme())
 	p.WithMask(image.Rect(0, 0, 10, 10), func(p *Painter) {
@@ -39,7 +38,7 @@ func TestMask_Full(t *testing.T) {
 }
 
 func TestMask_Inset(t *testing.T) {
-	surface := newTestSurface(10, 10)
+	surface := NewTestSurface(10, 10)
 
 	p := NewPainter(surface, NewTheme())
 	p.WithMask(image.Rect(0, 0, 10, 10), func(p *Painter) {
@@ -71,7 +70,7 @@ func TestMask_Inset(t *testing.T) {
 }
 
 func TestMask_FirstCell(t *testing.T) {
-	surface := newTestSurface(10, 10)
+	surface := NewTestSurface(10, 10)
 
 	p := NewPainter(surface, NewTheme())
 	p.WithMask(image.Rect(0, 0, 10, 10), func(p *Painter) {
@@ -103,7 +102,7 @@ func TestMask_FirstCell(t *testing.T) {
 }
 
 func TestMask_LastCell(t *testing.T) {
-	surface := newTestSurface(10, 10)
+	surface := NewTestSurface(10, 10)
 
 	p := NewPainter(surface, NewTheme())
 	p.WithMask(image.Rect(0, 0, 10, 10), func(p *Painter) {
@@ -135,7 +134,7 @@ func TestMask_LastCell(t *testing.T) {
 }
 
 func TestMask_MaskWithinEmptyMaskIsHidden(t *testing.T) {
-	surface := newTestSurface(10, 10)
+	surface := NewTestSurface(10, 10)
 
 	p := NewPainter(surface, NewTheme())
 	p.WithMask(image.Rect(0, 0, 0, 0), func(p *Painter) {
@@ -292,111 +291,3 @@ func TestWithStyle_Stacks(t *testing.T) {
 	}
 }
 
-type testCell struct {
-	Rune  rune
-	Style Style
-}
-
-type testSurface struct {
-	cells   map[image.Point]testCell
-	cursor  image.Point
-	size    image.Point
-	emptyCh rune
-}
-
-func newTestSurface(w, h int) *testSurface {
-	return &testSurface{
-		cells:   make(map[image.Point]testCell),
-		size:    image.Point{w, h},
-		emptyCh: '.',
-	}
-}
-
-func (s *testSurface) SetCell(x, y int, ch rune, style Style) {
-	s.cells[image.Point{x, y}] = testCell{
-		Rune:  ch,
-		Style: style,
-	}
-}
-
-func (s *testSurface) SetCursor(x, y int) {
-	s.cursor = image.Point{x, y}
-}
-
-func (s *testSurface) HideCursor() {
-	s.cursor = image.Point{}
-}
-
-func (s *testSurface) Begin() {
-	s.cells = make(map[image.Point]testCell)
-}
-
-func (s *testSurface) End() {
-	// NOP
-}
-
-func (s *testSurface) Size() image.Point {
-	return s.size
-}
-
-// String writes the testSurface's characters as a string.
-func (s *testSurface) String() string {
-	var buf bytes.Buffer
-	buf.WriteRune('\n')
-	for j := 0; j < s.size.Y; j++ {
-		for i := 0; i < s.size.X; i++ {
-			if cell, ok := s.cells[image.Point{i, j}]; ok {
-				buf.WriteRune(cell.Rune)
-				if w := runeWidth(cell.Rune); w > 1 {
-					i += w - 1
-				}
-			} else {
-				buf.WriteRune(s.emptyCh)
-			}
-		}
-		buf.WriteRune('\n')
-	}
-	return buf.String()
-}
-
-// FgColors renders the testSurface's foreground colors, using the digit 0-7 for painted cells.
-func (s *testSurface) FgColors() string {
-	var buf bytes.Buffer
-	buf.WriteRune('\n')
-	for j := 0; j < s.size.Y; j++ {
-		for i := 0; i < s.size.X; i++ {
-			if cell, ok := s.cells[image.Point{i, j}]; ok {
-				color := cell.Style.Fg
-				if cell.Style.Reverse {
-					color = cell.Style.Bg
-				}
-				buf.WriteRune('0' + rune(color))
-			} else {
-				buf.WriteRune(s.emptyCh)
-			}
-		}
-		buf.WriteRune('\n')
-	}
-	return buf.String()
-}
-
-// BgColors renders the testSurface's background colors, using the digit 0-7 for painted cells.
-func (s *testSurface) BgColors() string {
-	var buf bytes.Buffer
-	buf.WriteRune('\n')
-	for j := 0; j < s.size.Y; j++ {
-		for i := 0; i < s.size.X; i++ {
-			if cell, ok := s.cells[image.Point{i, j}]; ok {
-				color := cell.Style.Bg
-				if cell.Style.Reverse {
-					color = cell.Style.Fg
-				}
-				buf.WriteRune('0' + rune(color))
-			} else {
-				buf.WriteRune(s.emptyCh)
-			}
-		}
-		buf.WriteRune('\n')
-	}
-	return buf.String()
-}
